source-bigquery-batch: Move discovery into a separate file
